#6-refactor:Moved user CRUD operations to administration and implemented basic auth
